In an effort to keep everything going and not spend too long processing host commands when the baud rate is set to 1.5M... we only do one command per loop (i.e. the return at the end of the while loop if we processed a command)
And in an effort to reduce the amount of work we have to do, don't deque everything in the queue... only deque what is needed to be able to evaluate a complete command.  This could have an impact on reliable command receiption on the 232 interface (but it shouldn't unless they are sending commands without waiting for ACK --- I was doing that in an effort to "flood the CAN bus" with activity.... I didn't see any problems.  This change actually made things work better... but it seems more testing is in order).